+++ src/components/Select/types/Select.types.ts
@@ -1,25 +1,37 @@
-import { PropsWithoutChildren, Size } from '../../../types';
-import { Props as ReactSelectProps } from 'react-select';
+import { Size } from '../../../types';
+import { ReactNode, ReactText, RefObject } from 'react';
+import { ElementProps } from '../../../types/utils';
 
-export type StyledSelectProps = {
-    isMenuOpened?: Readonly<boolean>;
-    isFocused?: Readonly<boolean>;
-    isMulti?: Readonly<boolean>;
-    hasValue?: Readonly<boolean>;
-    isFullWidth?: Readonly<boolean>;
-    isDisabled?: Readonly<boolean>;
-    size?: Readonly<Size>;
-};
-
-export interface OptionType<T = string> {
-    value: T;
-    label: string;
+export interface OptionType<V = ReactText, L = ReactNode> {
+    value: V;
+    label: L;
 }
+
+type SharedSelectProps<T extends OptionType = OptionType> = {
+    options?: T[];
+    onOptionClick?: (option: T) => void;
+    isOpen?: boolean;
+    onClick?: ElementProps<HTMLDivElement>['onClick'];
+    isDisabled?: boolean;
+};
 
 export type ControllerSelectProps<
     T extends OptionType = OptionType
-> = ReactSelectProps<T>;
+> = SharedSelectProps<T> & {
+    defaultValue?: T;
+};
 
-export type SelectProps<
-    T extends OptionType = OptionType
-> = PropsWithoutChildren<StyledSelectProps & ControllerSelectProps<T>>;
+export type StyledComponentSelectProps = {
+    size?: Size;
+    visibleRows?: number;
+    isFullWidth?: boolean;
+};
+
+export type SelectProps<T extends OptionType = OptionType> = SharedSelectProps<
+    T
+> & {
+    value?: T;
+    placeholder?: T['label'];
+    ref?: RefObject<HTMLDivElement>;
+} & StyledComponentSelectProps &
+    ElementProps<HTMLDivElement>;