+++ src/components/Tooltips/constants/DEFAULT_POSITION.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_POSITION = 'right';