+++ src/components/Tooltips/types/CSSPosition.tsx
@@ -1,0 +1,1 @@
+export type CSSPosition = 'left' | 'right' | 'top' | 'bottom';