+++ src/components/Tooltips/utils/cleanupListeners.tsx
@@ -1,0 +1,10 @@
+export const cleanupListeners = (
+    currentEls: NodeListOf<HTMLElement>,
+    attachTooltip: (e: Event) => void,
+    cleanupTooltip: () => void
+) => {
+    for (const el of currentEls) {
+        el.removeEventListener('mouseenter', attachTooltip);
+        el.removeEventListener('mouseout', cleanupTooltip);
+    }
+};