+++ src/components/Transfer/components/LeftSide/styles/LeftSide.style.ts
@@ -1,6 +1,6 @@
 import styled, { css } from 'styled-components';
 import { StyledCheckboxLabel } from '../../../../Checkbox/view/alpha/style';
-import { StyledText } from '../../../../Typography/Text/Text.style';
+import { StyledText } from '../../../../Typography/Text/style';
 import { TransferProps } from '../../../types';
 import { getTransferSideHeight } from '../../../utils/getTransferSideHeight';
 import { StyledInputContainer } from '../../../../Input/view/composed/withFix/style';