+++ src/components/Transfer/hooks/useTransfer.tsx
@@ -3,10 +3,10 @@ import { ControllerTransferProps, TransferProps } from '../types';
 import { OptionType } from '../../Select';
 import React, { useRef, useState } from 'react';
 import { ButtonProps } from '../../Button';
-import { Text } from '../../Typography/Text';
-import { useOutsideClick } from './useOutsideClick';
+import { Text } from '../../Typography';
+import { useOutsideClick } from '../../../hooks/useOutsideClick';
 
-export const useTransfer = <T extends OptionType = OptionType>({
+export const useTransfer = <T extends OptionType<string, string>>({
     options,
     defaultValue,
     isDisabled,
@@ -86,6 +86,7 @@ export const useTransfer = <T extends OptionType = OptionType>({
     const sortedValue = sortOptions(value);
 
     return {
+        ...args,
         options: sortedOptions,
         value: sortedValue,
         onOptionClick,
@@ -125,7 +126,7 @@ export const useTransfer = <T extends OptionType = OptionType>({
     };
 };
 
-const sortOptions = <T extends OptionType = OptionType>(options: T[]) =>
+const sortOptions = <T extends OptionType<string, string>>(options: T[]) =>
     options.sort((a, b) =>
         a.label.toString().localeCompare(b.label.toString())
     );