+++ stories/Select.story.tsx
@@ -1,17 +1,16 @@
 import React from 'react';
-import { boolean, select, text, withKnobs } from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react';
-import { Select } from '../src/components/Select/view';
-import { useSelectChange } from '../src/components/Select/hooks';
-import { defaultValues } from './constants';
-import { Size } from '../src/types';
-import { action } from '@storybook/addon-actions';
 import {
     ControllerSelectProps,
+    Icon,
+    OptionType,
+    Select,
     SelectProps,
-    StyledSelectProps,
-} from '../src/components/Select/types';
-import { wrap } from '../src';
+    Size,
+    useSelect,
+} from '../src';
+import { boolean, select, withKnobs } from '@storybook/addon-knobs';
+import { defaultValues } from './constants';
 
 const stories = storiesOf('Select', module);
 
@@ -20,27 +19,45 @@ stories.addDecorator(withKnobs);
 stories.add(
     'Overview',
     () => {
-        const options = [
-            { label: 'Label 1', value: 'Value 1' },
-            { label: 'Label 2', value: 'Value 2' },
-            { label: 'Label 3', value: 'Value 3' },
-        ];
+        const iconHandlers = useSelect<OptionType<number>>({
+            options: [
+                {
+                    value: 1,
+                    label: <Icon name={`flagCzech`} />,
+                },
+                {
+                    value: 2,
+                    label: <Icon name={`flagSlovakia`} />,
+                },
+                {
+                    value: 3,
+                    label: <Icon name={`flagEU`} />,
+                },
+            ],
+            onOptionClick: console.log,
+        });
 
-        const SmallSelect = wrap(Select, useSelectChange);
-        SmallSelect.displayName = 'Select';
-        const MediumSelect = wrap(Select, useSelectChange);
-        MediumSelect.displayName = 'Select';
-        const MultiSelect = wrap(Select, useSelectChange);
-        MultiSelect.displayName = 'Select';
+        const textHandlers = useSelect<OptionType<number>>({
+            options: [
+                {
+                    value: 1,
+                    label: 'Česká Republika',
+                },
+                {
+                    value: 2,
+                    label: 'Slovenská Republika',
+                },
+                {
+                    value: 3,
+                    label: 'Evropa',
+                },
+            ],
+        });
 
         return (
             <>
-                <SmallSelect options={options} size={'small'} isMenuOpened />
-                <br />
-                <MediumSelect options={options} isFullWidth={true} />
-                <br />
-                <MultiSelect options={options} isMulti />
-                <br />
+                <Select isFullWidth size={'large'} {...iconHandlers} />
+                <Select {...textHandlers} />
             </>
         );
     },
@@ -52,60 +69,51 @@ stories.add(
 stories.add(
     'Playground',
     () => {
-        const options = [
-            { label: 'Label 1', value: 'Value 1' },
-            { label: 'Label 2', value: 'Value 2' },
-            { label: 'Label 3', value: 'Value 3' },
-        ];
-
-        const isDisabled: SelectProps['isDisabled'] = boolean(
-            `isDisabled:`,
-            defaultValues.isDisabled
-        );
-
         const isFullWidth: SelectProps['isFullWidth'] = boolean(
             `isFullWidth:`,
             defaultValues.isFullWidth
         );
-
-        const isMulti: SelectProps['isMulti'] = boolean(
-            `isMulti:`,
-            defaultValues.isMulti
-        );
-
-        const isFocused: SelectProps['isFocused'] = boolean(
-            `isFocused:`,
-            false
-        );
-
         const size: Size = select(
             `size:`,
             ['small', `medium`, 'large'],
             defaultValues.size
         ) as Size;
 
-        const placeholder: SelectProps['placeholder'] = text(
-            `placeholder`,
-            defaultValues.placeholder
+        const selectProps: ControllerSelectProps<OptionType<number>> = {
+            options: [
+                {
+                    value: 1,
+                    label: 'Label 1',
+                },
+                {
+                    value: 2,
+                    label: 'Label 2',
+                },
+                {
+                    value: 3,
+                    label: 'Label 3',
+                },
+                {
+                    value: 4,
+                    label:
+                        'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx',
+                },
+            ],
+        };
+
+        const demoSelectHandlers = useSelect<OptionType<number>>({
+            ...selectProps,
+        });
+
+        return (
+            <div>
+                <Select
+                    isFullWidth={isFullWidth}
+                    size={size}
+                    {...demoSelectHandlers}
+                />
+            </div>
         );
-
-        const onChange: SelectProps['onChange'] = action(`onChange`);
-
-        const styledProps: StyledSelectProps = {
-            isDisabled,
-            isFullWidth,
-            isMulti,
-            isFocused,
-            size,
-        };
-
-        const controlProps: ControllerSelectProps = {
-            onChange,
-            placeholder,
-            options,
-        };
-
-        return <Select {...styledProps} {...controlProps} />;
     },
     {
         info: { inline: true },