+++ stories/style/index.ts
@@ -1,1 +1,2 @@
+export * from './Icon.style';
 export * from './Stories.style';